requirements.txt
# Core Dependencies
pymongo==4.15.3
openai==2.6.1
python-dotenv==1.2.1

# LangGraph and LangChain
langgraph==1.0.1
langchain-core==1.0.1

# Memory and Embeddings
chromadb==1.3.0
sentence-transformers==5.1.2

# FastAPI and Web
fastapi==0.120.1
uvicorn[standard]==0.38.0
websockets==15.0.1

# Data Processing
pydantic==2.12.3

# Utilities
python-multipart==0.0.20
